FAT: some refactoring to make the FAT area more extensible.
